refactor: fixed a bug in signin page and added a scrollview in schedule washing
